채팅기능 구현 및 엔티티 속성 이름 통일

+++ src/main/java/com/mutsa/springboot_auction/domain/auctionCategory/AuctionCategoryId.java
@@ -11,6 +11,6 @@ import lombok.*;
 @EqualsAndHashCode
 public class AuctionCategoryId {
     // 복합키 사용을 위한 클래스
-    private Integer auctionId; // FK
-    private Integer categoryId; // FK
+    private Long auctionId; // FK
+    private Long categoryId; // FK
 }
